# [hotfix-to-7.5.6] raftstore: remove stale ranges by DeleteByKeys rather than ingesting. (#18040)