This is a manual cherry-pick of #18040

<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--
Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.
-->

Issue Number: Close #18107, Ref https://github.com/tikv/tikv/issues/18042

<!--
You could use "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#format-of-the-commit-message.
-->
This PR mainly contains the following parts for optimization on scaling, used to mitigate the impacts, introduced by unnecessary ingesting sst files:
- Directly clearing stale ranges by DeleteByKeys during the balancing regions process.
- Do not clear data of offline stores during the scale-in process, as this data will be automatically cleared when the corresponding node goes offline.

What's Changed:

```commit-message
Optimizing the processing of clearing stale-ranges by DeleteByKeys, rather than ingesting.

This optimization is used to mitigate the impacts, introduced by unnecessary ingesting sst
files on latency when scaling.
```

### Related changes

- [ ] PR to update `pingcap/docs`/`pingcap/docs-cn`:
- [ ] Need to cherry-pick to the release branch

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Taking the `grpc messages duration` in TiKV metrics panel as examples, positive performance feedbacks on reducing the long-tail latency can be reviewed from the following comparison results:
| Commit | Comparisons on `grpc messages duration` |
| --- | --- |
| master | ![image](https://github.com/user-attachments/assets/38b158a1-e0c3-43b5-9c0f-f40517d9db67) |
| This PR | ![image](https://github.com/user-attachments/assets/2e9f45b8-d6ac-4db4-9a1c-abb6463e883c) |

Meanwhile, the following E2E long-tail reduction also proves that this PR makes positive improvements effects:
| Commit | Scale-Out (Before -> Scaling) | Scale-in (Before -> Scaling) |
| --- | --- | --- |
| master | P99: 7.2ms -> ~ 11.3 ms  P999: 62ms -> ~89.2 ms | P99: 7.2ms -> ~ 16.1 ms  P999: 62ms -> ~100 ms |
| This PR | P99: 7.2 ms -> ~10.6 ms   P999: 62 ms -> ~76.9ms | P99: 7.2ms -> ~ 15.3 ms   P999: 62ms -> ~96 ms |

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

### Release note
<!-- 
Compatibility change, improvement, bugfix, and new feature need a release note.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

If you don't think this PR needs a release note then fill it with None.
If this PR will be picked to release branch, then a release note is probably required.
-->

```release-note
Optimizing the processing of clearing stale-ranges by DeleteByKeys to mitigate the impacts
on latency.
```